Moving TemporaryS3 to saws-s3
