+++ saws-s3/src/main/scala/com/ambiata/saws/s3/TemporaryS3.scala
@@ -1,11 +1,10 @@
-package com.ambiata.saws.testing
+package com.ambiata.saws.s3
 
 import java.util.UUID
 
 import com.ambiata.mundane.control._
 import com.ambiata.saws.core._
-import com.ambiata.saws.s3._
-import com.ambiata.saws.testing.TemporaryS3._
+import com.ambiata.saws.s3.TemporaryS3._
 
 import scalaz._, Scalaz._, effect._
